# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "EntertainmentMp3 (video and News App) Api",
    "description": "API for educational purposes Blog OpenSource to build an android app in my 100Days of Code by <b>Bra Koose's 404 Solutions </b> using NodeJs."
  },
  "host": "localhost:5000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Authorization": {
      "type": "apikey",
      "name": "Authorization",
      "description": "Value Bearer",
      "in": "header",
      "scheme": "bearer"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "scheme": {
              "$ref": "#/definitions/LoginModel"
            },
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterModel"
            }
          }
        ],
        "responses": {
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/auth/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VerifyEmailModel"
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/forgotPassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ForgotPassWordModel"
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/resetPassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ResetPasswordModel"
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/changePassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChangePasswordModel"
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "LoginModel": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "admin@entertainmentmp3.com"
        },
        "password": {
          "type": "string",
          "example": "DoBetter$$$2022"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "RegisterModel": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Bra Koose"
        },
        "email": {
          "type": "string",
          "example": "yourmail@whatevermail.com"
        },
        "password": {
          "type": "string",
          "example": "AnyPassword123YouWantButNotThis"
        }
      },
      "required": [
        "name",
        "email",
        "password"
      ]
    },
...
    "VerifyEmailModel": {
      "type": "object",
      "properties": {
        "code": {
          "type": "number",
          "example": 909090
        }
      },
      "required": [
        "code"
      ]
    },
    "ChangePasswordModel": {
      "type": "object",
      "properties": {
        "oldPassword": {
          "type": "string",
          "example": "YourPassword123"
        },
        "newPassword": {
          "type": "string",
          "example": "NewPassword#9090"
        }
      },
      "required": [
        "oldPassword",
        "newPassword"
      ]
    },
...
    "ResetPasswordModel": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "koose@gmail.com"
        },
        "code": {
          "type": "number",
          "example": 9999999
        },
        "newPassword": {
          "type": "string",
          "example": "HeyPassword90@"
        }
      },
      "required": [
        "email",
        "code",
        "newPassword"
      ]
    }
  }
}